--- app/(temp)/_layout.jsx
@@ -1,20 +1,0 @@
-import React from "react";
-import { Stack } from "expo-router";
-
-const tempLayout = () => {
-  return (
-    <>
-      <Stack>
-        <Stack.Screen
-          name="home"
-          options={{
-            headerShown: false,
-          }}
-        />
-
-      </Stack>
-    </>
-  );
-};
-
-export default tempLayout;
